Detect official eft game install
